remove some old-style C casts

+++ source/particle/generator/uniform_radial.cc
@@ -95,7 +95,10 @@ namespace aspect
             else if (dim == 3)
               {
                 const unsigned int theta_particles = round(sqrt(particles_per_layer[i]));
-                const unsigned int phi_particles = round ((double) particles_per_layer[i] / (double) theta_particles);
+                const unsigned int phi_particles = round(
+                                                     static_cast<double>(particles_per_layer[i])
+                                                     /
+                                                     static_cast<double>(theta_particles));
                 const double theta_spacing = (P_max[2] - P_min[2]) / fmax(theta_particles-1,1);
 
                 for (unsigned int j = 0; j < theta_particles; ++j)
@@ -103,7 +106,7 @@ namespace aspect
                     spherical_coordinates[2] = P_min[2] + j * theta_spacing;
 
                     // Average value of sin(n) from 0 to 180 degrees is (2/pi)
-                    const unsigned int adjusted_phi_particles = std::max(static_cast<unsigned int> (phi_particles * std::sin(spherical_coordinates[2])),(unsigned int) 1);
+                    const unsigned int adjusted_phi_particles = std::max(static_cast<unsigned int> (phi_particles * std::sin(spherical_coordinates[2])), 1u);
                     const double phi_spacing = (P_max[1] - P_min[1]) / fmax(adjusted_phi_particles-1,1);
                     for (unsigned int k = 0; k < adjusted_phi_particles; ++k)
                       {
